+++ zorphy/lib/src/generators/json_generator.dart
@@ -98,7 +98,7 @@ class JsonGenerator extends UniversalGenerator {
             final info = manualField.first.jsonKeyInfo!;
             final jsonFieldName = _escapeDartStringLiteral(info.name ?? f.name);
             bodyLines.add(
-                "      ${f.name}: json['$jsonFieldName'] != null ? ${info.fromJson}(json['$jsonFieldName'] as Map<String, dynamic>) as ${f.type} : null,");
+                "      ${f.name}: json['$jsonFieldName'] != null ? ${info.fromJson}(json['$jsonFieldName'] as Map<String, dynamic>) as ${f.type!.replaceAll(r'$', '')} : null,");
           } else {
             bodyLines.add('      ${f.name}: instance.${f.name},');
           }
@@ -270,7 +270,7 @@ class JsonGenerator extends UniversalGenerator {
           final info = manualField.first.jsonKeyInfo!;
           final jsonFieldName = _escapeDartStringLiteral(info.name ?? f.name);
           bodyLines.add(
-              "      ${f.name}: json['$jsonFieldName'] != null ? ${info.fromJson}(json['$jsonFieldName'] as Map<String, dynamic>) as ${f.type} : null,");
+              "      ${f.name}: json['$jsonFieldName'] != null ? ${info.fromJson}(json['$jsonFieldName'] as Map<String, dynamic>) as ${f.type!.replaceAll(r'$', '')} : null,");
         } else {
           bodyLines.add('      ${f.name}: instance.${f.name},');
         }
@@ -328,7 +328,8 @@ class JsonGenerator extends UniversalGenerator {
         body.add(
             "if (${f.name} != null) data['$jsonFieldName'] = ${info.toJson}(${f.name}!);");
       }
-      body.add('return _sanitizeJson(data);');
+      body.add('_sanitizeJson(data);');
+      body.add('return data;');
       specs.add(Method((m) {
         m.name = 'toJsonLean';
         m.returns = referType('Map<String, dynamic>');
@@ -347,7 +348,8 @@ class JsonGenerator extends UniversalGenerator {
         body.add(
             "if (${f.name} != null) data['$jsonFieldName'] = ${info.toJson}(${f.name}!);");
       }
-      body.add('return _sanitizeJson(data);');
+      body.add('_sanitizeJson(data);');
+      body.add('return data;');
       specs.add(Method((m) {
         m.name = 'toJsonLean';
         m.returns = referType('Map<String, dynamic>');