+++ zorphy_migrator/test/fixtures/exchangeable_object_assert_body/expected.dart
@@ -1,0 +1,9 @@
+@Zorphy(
+  kind: ZorphyKind.valueObject,
+  generateJson: true,
+  generateCompareTo: true,
+)
+abstract class $UIImage {
+  ///The name of the image asset.
+  String? get name;
+}